slight reorg and simplification
