+++ controllers/kokumetricsconfig_controller.go
@@ -7,11 +7,9 @@ package controllers
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"math/rand"
 	"os"
-	"path/filepath"
 	"reflect"
 	"strings"
 	"time"
@@ -19,7 +17,6 @@ import (
 	gologr "github.com/go-logr/logr"
 	promv1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
@@ -67,29 +64,27 @@ type MetricsConfigReconciler struct {
 	InCluster bool
 	Namespace string
 
-	cvClientBuilder               cv.ClusterVersionBuilder
 	promCollector                 *collector.PrometheusCollector
 	disablePreviousDataCollection bool
 	overrideSecretPath            bool
 }
 
-type previousAuthValidation struct {
-	secretName string
-	username   string
-	password   string
-	err        error
-	timestamp  metav1.Time
+// GetClientset returns a clientset based on rest.config
+func GetClientset() (*kubernetes.Clientset, error) {
+	config, err := config.GetConfig()
+	if err != nil {
+		return nil, err
+	}
+	// creates the clientset
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return nil, err
+	}
+	return clientset, nil
 }
 
-type serializedAuthMap struct {
-	Auths map[string]serializedAuth `json:"auths"`
-}
-type serializedAuth struct {
-	Auth string `json:"auth"`
-}
-
-// StringReflectSpec Determine if the string Status item reflects the Spec item if not empty, otherwise take the default value.
-func StringReflectSpec(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig, specItem *string, statusItem *string, defaultVal string) (string, bool) {
+// stringReflectSpec Determine if the string Status item reflects the Spec item if not empty, otherwise take the default value.
+func stringReflectSpec(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig, specItem *string, statusItem *string, defaultVal string) (string, bool) {
 	// Update statusItem if needed
 	changed := false
 	if *statusItem == "" || !reflect.DeepEqual(*specItem, *statusItem) {
@@ -106,18 +101,18 @@ func StringReflectSpec(r *MetricsConfigReconciler, cr *metricscfgv1beta1.Metrics
 	return *statusItem, changed
 }
 
-// ReflectSpec Determine if the Status item reflects the Spec item if not empty, otherwise set a default value if applicable.
-func ReflectSpec(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig) {
+// reflectSpec Determine if the Status item reflects the Spec item if not empty, otherwise set a default value if applicable.
+func reflectSpec(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig) {
 
-	StringReflectSpec(r, cr, &cr.Spec.APIURL, &cr.Status.APIURL, metricscfgv1beta1.DefaultAPIURL)
-	StringReflectSpec(r, cr, &cr.Spec.Authentication.AuthenticationSecretName, &cr.Status.Authentication.AuthenticationSecretName, "")
+	stringReflectSpec(r, cr, &cr.Spec.APIURL, &cr.Status.APIURL, metricscfgv1beta1.DefaultAPIURL)
+	stringReflectSpec(r, cr, &cr.Spec.Authentication.AuthenticationSecretName, &cr.Status.Authentication.AuthenticationSecretName, "")
 
 	if !reflect.DeepEqual(cr.Spec.Authentication.AuthType, cr.Status.Authentication.AuthType) {
 		cr.Status.Authentication.AuthType = cr.Spec.Authentication.AuthType
 	}
 	cr.Status.Upload.ValidateCert = cr.Spec.Upload.ValidateCert
 
-	StringReflectSpec(r, cr, &cr.Spec.Upload.IngressAPIPath, &cr.Status.Upload.IngressAPIPath, metricscfgv1beta1.DefaultIngressPath)
+	stringReflectSpec(r, cr, &cr.Spec.Upload.IngressAPIPath, &cr.Status.Upload.IngressAPIPath, metricscfgv1beta1.DefaultIngressPath)
 	cr.Status.Upload.UploadToggle = cr.Spec.Upload.UploadToggle
 
 	// set the default max file size for packaging
@@ -140,8 +135,8 @@ func ReflectSpec(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig
 		cr.Status.Upload.UploadCycle = cr.Spec.Upload.UploadCycle
 	}
 
-	StringReflectSpec(r, cr, &cr.Spec.Source.SourcesAPIPath, &cr.Status.Source.SourcesAPIPath, metricscfgv1beta1.DefaultSourcesPath)
-	StringReflectSpec(r, cr, &cr.Spec.Source.SourceName, &cr.Status.Source.SourceName, "")
+	stringReflectSpec(r, cr, &cr.Spec.Source.SourcesAPIPath, &cr.Status.Source.SourcesAPIPath, metricscfgv1beta1.DefaultSourcesPath)
+	stringReflectSpec(r, cr, &cr.Spec.Source.SourceName, &cr.Status.Source.SourceName, "")
 
 	cr.Status.Source.CreateSource = cr.Spec.Source.CreateSource
 
@@ -149,139 +144,44 @@ func ReflectSpec(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig
 		cr.Status.Source.CheckCycle = cr.Spec.Source.CheckCycle
 	}
 
-	StringReflectSpec(r, cr, &cr.Spec.PrometheusConfig.SvcAddress, &cr.Status.Prometheus.SvcAddress, metricscfgv1beta1.DefaultPrometheusSvcAddress)
+	stringReflectSpec(r, cr, &cr.Spec.PrometheusConfig.SvcAddress, &cr.Status.Prometheus.SvcAddress, metricscfgv1beta1.DefaultPrometheusSvcAddress)
 	cr.Status.Prometheus.SkipTLSVerification = cr.Spec.PrometheusConfig.SkipTLSVerification
 	cr.Status.Prometheus.ContextTimeout = cr.Spec.PrometheusConfig.ContextTimeout
 	cr.Status.Prometheus.DisabledMetricsCollectionCostManagement = cr.Spec.PrometheusConfig.DisableMetricsCollectionCostManagement
 	cr.Status.Prometheus.DisabledMetricsCollectionResourceOptimization = cr.Spec.PrometheusConfig.DisableMetricsCollectionResourceOptimization
 }
 
-// GetClientset returns a clientset based on rest.config
-func GetClientset() (*kubernetes.Clientset, error) {
-	config, err := config.GetConfig()
-	if err != nil {
-		return nil, err
+func setClusterID(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig) error {
+	if cr.Status.ClusterID != "" && cr.Status.ClusterVersion != "" {
+		return nil
 	}
-	// creates the clientset
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-	return clientset, nil
-}
 
-// GetClusterID Collects the cluster identifier and version from the Cluster Version custom resource object
-func GetClusterID(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig) error {
-	log := log.WithName("GetClusterID")
-	// Get current ClusterVersion
-	cvClient := r.cvClientBuilder.New(r.Client)
+	cvClient := cv.NewCVClient(r.Client)
 	clusterVersion, err := cvClient.GetClusterVersion()
 	if err != nil {
 		return err
 	}
+
 	log.Info("cluster version found", "ClusterVersion", clusterVersion.Spec)
 	if clusterVersion.Spec.ClusterID != "" {
 		cr.Status.ClusterID = string(clusterVersion.Spec.ClusterID)
 	}
-	if clusterVersion.Spec.Channel != "" {
-		cr.Status.ClusterVersion = string(clusterVersion.Spec.Channel)
+	if clusterVersion.Status.Desired.Version != "" {
+		cr.Status.ClusterVersion = string(clusterVersion.Status.Desired.Version)
 	}
 	return nil
 }
 
-// GetPullSecretToken Obtain the bearer token string from the pull secret in the openshift-config namespace
-func GetPullSecretToken(r *MetricsConfigReconciler, authConfig *crhchttp.AuthConfig) error {
-	ctx := context.Background()
-	log := log.WithName("GetPullSecretToken")
-
-	secret, err := r.Clientset.CoreV1().Secrets(openShiftConfigNamespace).Get(ctx, pullSecretName, metav1.GetOptions{})
-	if err != nil {
-		switch {
-		case errors.IsNotFound(err):
-			log.Error(err, "pull-secret does not exist")
-		case errors.IsForbidden(err):
-			log.Error(err, "operator does not have permission to check pull-secret")
-		default:
-			log.Error(err, "could not check pull-secret")
-		}
-		return err
-	}
-
-	tokenFound := false
-	encodedPullSecret := secret.Data[pullSecretDataKey]
-	if len(encodedPullSecret) <= 0 {
-		return fmt.Errorf("cluster authorization secret did not have data")
-	}
-	var pullSecret serializedAuthMap
-	if err := json.Unmarshal(encodedPullSecret, &pullSecret); err != nil {
-		log.Error(err, "unable to unmarshal cluster pull-secret")
-		return err
-	}
-	if auth, ok := pullSecret.Auths[pullSecretAuthKey]; ok {
-		token := strings.TrimSpace(auth.Auth)
-		if strings.Contains(token, "\n") || strings.Contains(token, "\r") {
-			return fmt.Errorf("cluster authorization token is not valid: contains newlines")
-		}
-		if len(token) > 0 {
-			log.Info("found cloud.openshift.com token")
-			authConfig.BearerTokenString = token
-			tokenFound = true
-		} else {
-			return fmt.Errorf("cluster authorization token is not found")
-		}
-	} else {
-		return fmt.Errorf("cluster authorization token was not found in secret data")
-	}
-	if !tokenFound {
-		return fmt.Errorf("cluster authorization token is not found")
-	}
-	return nil
-}
-
-// GetAuthSecret Obtain the username and password from the authentication secret provided in the current namespace
-func GetAuthSecret(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig, authConfig *crhchttp.AuthConfig, reqNamespace types.NamespacedName) error {
-	ctx := context.Background()
-	log := log.WithName("GetAuthSecret")
-
-	if previousValidation == nil || previousValidation.secretName != cr.Status.Authentication.AuthenticationSecretName {
-		previousValidation = &previousAuthValidation{secretName: cr.Status.Authentication.AuthenticationSecretName}
-	}
-
-	log.Info("secret namespace", "namespace", reqNamespace.Namespace)
-	secret := &corev1.Secret{}
-	namespace := types.NamespacedName{
-		Namespace: reqNamespace.Namespace,
-		Name:      cr.Status.Authentication.AuthenticationSecretName}
-	err := r.Get(ctx, namespace, secret)
-	if err != nil {
-		switch {
-		case errors.IsNotFound(err):
-			log.Error(err, "secret does not exist")
-		case errors.IsForbidden(err):
-			log.Error(err, "operator does not have permission to check secret")
-		default:
-			log.Error(err, "could not check secret")
-		}
-		return err
-	}
-
-	keys := make(map[string]string)
-	for k, v := range secret.Data {
-		keys[strings.ToLower(k)] = string(v)
-	}
-
-	for _, k := range []string{authSecretUserKey, authSecretPasswordKey} {
-		if len(keys[k]) <= 0 {
-			msg := fmt.Sprintf("secret not found with expected %s data", k)
+func setOperatorCommit(r *MetricsConfigReconciler) {
+	log := log.WithName("setOperatorCommit")
+	if GitCommit == "" {
+		commit, exists := os.LookupEnv("GIT_COMMIT")
+		if exists {
+			msg := fmt.Sprintf("using git commit from environment: %s", commit)
 			log.Info(msg)
-			return fmt.Errorf(msg)
+			GitCommit = commit
 		}
 	}
-
-	authConfig.BasicAuthUser = keys[authSecretUserKey]
-	authConfig.BasicAuthPassword = keys[authSecretPasswordKey]
-
-	return nil
 }
 
 func checkCycle(log gologr.Logger, cycle int64, lastExecution metav1.Time, action string) bool {
@@ -302,104 +202,6 @@ func checkCycle(log gologr.Logger, cycle int64, lastExecution metav1.Time, actio
 	log.Info(fmt.Sprintf("not time to execute the %s", action))
 	return false
 
-}
-
-func setClusterID(r *MetricsConfigReconciler, cr *metricscfgv1beta1.MetricsConfig) error {
-	if cr.Status.ClusterID == "" || cr.Status.ClusterVersion == "" {
-		r.cvClientBuilder = cv.NewBuilder()
-		err := GetClusterID(r, cr)
-		return err
-	}
-	return nil
-}
-
-func setAuthentication(r *MetricsConfigReconciler, authConfig *crhchttp.AuthConfig, cr *metricscfgv1beta1.MetricsConfig, reqNamespace types.NamespacedName) error {
-	log := log.WithName("setAuthentication")
-	cr.Status.Authentication.AuthenticationCredentialsFound = &trueDef
-	if cr.Status.Authentication.AuthType == metricscfgv1beta1.Token {
-		cr.Status.Authentication.ValidBasicAuth = nil
-		cr.Status.Authentication.AuthErrorMessage = ""
-		cr.Status.Authentication.LastVerificationTime = nil
-		// Get token from pull secret
-		err := GetPullSecretToken(r, authConfig)
-		if err != nil {
-			log.Error(nil, "failed to obtain cluster authentication token")
-			cr.Status.Authentication.AuthenticationCredentialsFound = &falseDef
-			cr.Status.Authentication.AuthErrorMessage = err.Error()
-		}
-		return err
-	} else if cr.Spec.Authentication.AuthenticationSecretName != "" {
-		// Get user and password from auth secret in namespace
-		err := GetAuthSecret(r, cr, authConfig, reqNamespace)
-		if err != nil {
-			log.Error(nil, "failed to obtain authentication secret credentials")
-			cr.Status.Authentication.AuthenticationCredentialsFound = &falseDef
-			cr.Status.Authentication.AuthErrorMessage = err.Error()
-			cr.Status.Authentication.ValidBasicAuth = &falseDef
-		}
-		return err
-	} else {
-		// No authentication secret name set when using basic auth
-		cr.Status.Authentication.AuthenticationCredentialsFound = &falseDef
-		err := fmt.Errorf("no authentication secret name set when using basic auth")
-		cr.Status.Authentication.AuthErrorMessage = err.Error()
-		cr.Status.Authentication.ValidBasicAuth = &falseDef
-		return err
-	}
-}
-
-func validateCredentials(r *MetricsConfigReconciler, handler *sources.SourceHandler, cr *metricscfgv1beta1.MetricsConfig, cycle int64) error {
-	log := log.WithName("validateCredentials")
-
-	if cr.Spec.Authentication.AuthType == metricscfgv1beta1.Token {
-		// no need to validate token auth
-		return nil
-	}
-
-	if previousValidation == nil {
-		previousValidation = &previousAuthValidation{}
-	}
-
-	if previousValidation.password == handler.Auth.BasicAuthPassword &&
-		previousValidation.username == handler.Auth.BasicAuthUser &&
-		!checkCycle(log, cycle, previousValidation.timestamp, "credential verification") {
-		return previousValidation.err
-	}
-
-	log.Info("validating credentials")
-	client := crhchttp.GetClient(handler.Auth)
-	_, err := sources.GetSources(handler, client)
-
-	previousValidation.username = handler.Auth.BasicAuthUser
-	previousValidation.password = handler.Auth.BasicAuthPassword
-	previousValidation.err = err
-	previousValidation.timestamp = metav1.Now()
-
-	cr.Status.Authentication.LastVerificationTime = &previousValidation.timestamp
-
-	if err != nil && strings.Contains(err.Error(), "401") {
-		msg := fmt.Sprintf("cloud.redhat.com credentials are invalid. Correct the username/password in `%s`. Updated credentials will be re-verified during the next reconciliation.", cr.Spec.Authentication.AuthenticationSecretName)
-		log.Info(msg)
-		cr.Status.Authentication.AuthErrorMessage = msg
-		cr.Status.Authentication.ValidBasicAuth = &falseDef
-		return err
-	}
-	log.Info("credentials are valid")
-	cr.Status.Authentication.AuthErrorMessage = ""
-	cr.Status.Authentication.ValidBasicAuth = &trueDef
-	return nil
-}
-
-func setOperatorCommit(r *MetricsConfigReconciler) {
-	log := log.WithName("setOperatorCommit")
-	if GitCommit == "" {
-		commit, exists := os.LookupEnv("GIT_COMMIT")
-		if exists {
-			msg := fmt.Sprintf("using git commit from environment: %s", commit)
-			log.Info(msg)
-			GitCommit = commit
-		}
-	}
 }
 
 func checkSource(r *MetricsConfigReconciler, handler *sources.SourceHandler, cr *metricscfgv1beta1.MetricsConfig) {
@@ -423,92 +225,6 @@ func checkSource(r *MetricsConfigReconciler, handler *sources.SourceHandler, cr 
 		cr.Status.Source.SourceDefined = &defined
 		cr.Status.Source.LastSourceCheckTime = lastCheck
 	}
-}
-
-func packageFiles(p *packaging.FilePackager) {
-	log := log.WithName("packageAndUpload")
-
-	// if its time to package
-	if !checkCycle(log, *p.CR.Status.Upload.UploadCycle, p.CR.Status.Packaging.LastSuccessfulPackagingTime, "file packaging") {
-		return
-	}
-
-	// Package and split the payload if necessary
-	p.CR.Status.Packaging.PackagingError = ""
-	if err := p.PackageReports(); err != nil {
-		log.Error(err, "PackageReports failed")
-		// update the CR packaging error status
-		p.CR.Status.Packaging.PackagingError = err.Error()
-	}
-}
-
-func uploadFiles(r *MetricsConfigReconciler, authConfig *crhchttp.AuthConfig, cr *metricscfgv1beta1.MetricsConfig, dirCfg *dirconfig.DirectoryConfig, packager *packaging.FilePackager) error {
-	log := log.WithName("uploadFiles")
-
-	// if its time to upload/package
-	if !*cr.Spec.Upload.UploadToggle {
-		log.Info("operator is configured to not upload reports")
-		return nil
-	}
-	if !checkCycle(log, *cr.Status.Upload.UploadCycle, cr.Status.Upload.LastSuccessfulUploadTime, "upload") {
-		return nil
-	}
-
-	uploadFiles, err := dirCfg.Upload.GetFiles()
-	if err != nil {
-		log.Error(err, "failed to read upload directory")
-		return err
-	}
-
-	if len(uploadFiles) <= 0 {
-		log.Info("no files to upload")
-		return nil
-	}
-
-	log.Info("files ready for upload: " + strings.Join(uploadFiles, ", "))
-	log.Info("pausing for " + fmt.Sprintf("%d", *cr.Status.Upload.UploadWait) + " seconds before uploading")
-	time.Sleep(time.Duration(*cr.Status.Upload.UploadWait) * time.Second)
-	for _, file := range uploadFiles {
-		if !strings.Contains(file, "tar.gz") {
-			continue
-		}
-
-		manifestInfo, err := packager.GetFileInfo(filepath.Join(dirCfg.Upload.Path, file))
-		if err != nil {
-			log.Error(err, "Could not read file information from tar.gz")
-			continue
-		}
-
-		log.Info(fmt.Sprintf("uploading file: %s", file))
-		// grab the body and the multipart file header
-		body, contentType, err := crhchttp.GetMultiPartBodyAndHeaders(filepath.Join(dirCfg.Upload.Path, file))
-		if err != nil {
-			log.Error(err, "failed to set multipart body and headers")
-			return err
-		}
-		ingressURL := cr.Status.APIURL + cr.Status.Upload.IngressAPIPath
-		uploadStatus, uploadTime, requestID, err := crhchttp.Upload(authConfig, contentType, "POST", ingressURL, body, manifestInfo, file)
-		cr.Status.Upload.LastUploadStatus = uploadStatus
-		cr.Status.Upload.LastPayloadName = file
-		cr.Status.Upload.LastPayloadFiles = manifestInfo.Files
-		cr.Status.Upload.LastPayloadManifestID = manifestInfo.UUID
-		cr.Status.Upload.LastPayloadRequestID = requestID
-		cr.Status.Upload.UploadError = ""
-		if err != nil {
-			log.Error(err, "upload failed")
-			cr.Status.Upload.UploadError = err.Error()
-			return nil
-		}
-		if strings.Contains(uploadStatus, "202") {
-			cr.Status.Upload.LastSuccessfulUploadTime = uploadTime
-			// remove the tar.gz after a successful upload
-			log.Info("removing tar file since upload was successful")
-			if err := os.Remove(filepath.Join(dirCfg.Upload.Path, file)); err != nil {
-				log.Error(err, "error removing tar file")
-			}
-		}
-	}
-	return nil
 }
 
 func configurePVC(r *MetricsConfigReconciler, req ctrl.Request, cr *metricscfgv1beta1.MetricsConfig) (*ctrl.Result, error) {
@@ -553,6 +269,19 @@ func configurePVC(r *MetricsConfigReconciler, req ctrl.Request, cr *metricscfgv1
 	return nil, nil
 }
 
+// concatErrs combines all the errors into one error
+func concatErrs(errors ...error) error {
+	var err error
+	var errstrings []string
+	for _, e := range errors {
+		errstrings = append(errstrings, e.Error())
+	}
+	if len(errstrings) > 0 {
+		err = fmt.Errorf(strings.Join(errstrings, "\n"))
+	}
+	return err
+}
+
 // +kubebuilder:rbac:groups=koku-metrics-cfg.openshift.io,namespace=koku-metrics-operator,resources=kokumetricsconfigs,verbs=get;list;watch;create;update;patch;delete
 // +kubebuilder:rbac:groups=koku-metrics-cfg.openshift.io,namespace=koku-metrics-operator,resources=kokumetricsconfigs/status,verbs=get;update;patch
 // +kubebuilder:rbac:groups=operators.coreos.com,namespace=koku-metrics-operator,resources=clusterserviceversions,verbs=get;list;watch;update;patch
@@ -581,7 +310,7 @@ func (r *MetricsConfigReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	log.Info("reconciling custom resource", "MetricsConfig", cr)
 
 	// reflect the spec values into status
-	ReflectSpec(r, cr)
+	reflectSpec(r, cr)
 
 	if r.InCluster {
 		res, err := configurePVC(r, req, cr)
@@ -753,17 +482,4 @@ func (r *MetricsConfigReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&metricscfgv1beta1.MetricsConfig{}).
 		Complete(r)
-}
-
-// concatErrs combines all the errors into one error
-func concatErrs(errors ...error) error {
-	var err error
-	var errstrings []string
-	for _, e := range errors {
-		errstrings = append(errstrings, e.Error())
-	}
-	if len(errstrings) > 0 {
-		err = fmt.Errorf(strings.Join(errstrings, "\n"))
-	}
-	return err
 }
